refactor: move trapping data charts to separate files
