+++ src/screens/trapping-data/components/line-chart/components/index.js
@@ -1,0 +1,5 @@
+import AvgProbChart from './avg-prob-chart';
+import SPBChart from './spb-chart';
+import TotalChart from './total-chart';
+
+export { AvgProbChart, SPBChart, TotalChart };
