+++ src/screens/trapping-data/components/line-chart/index.js
@@ -1,30 +1,3 @@
-import { connect } from 'react-redux';
-
 import LineChart from './component';
 
-const mapStateToProps = (state) => {
-  const {
-    data: {
-      yearData,
-    },
-    selections: {
-      startYear,
-      endYear,
-    },
-  } = state;
-
-  return {
-    yearData,
-    startYear,
-    endYear,
-  };
-};
-
-const mapDispatchToProps = (dispatch) => {
-  return {};
-};
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(LineChart);
+export default LineChart;
